move decoration to d
