+++ src/d.ts
@@ -36,6 +36,30 @@ export function isHNode(child: DNode): child is HNode {
 	return Boolean(child && (typeof child !== 'string') && child.type === HNODE);
 }
 
+export interface DecoratorPredicate {
+	(dNode: DNode): Boolean;
+}
+
+export interface DecoratorModifier {
+	(dNode: DNode): DNode;
+}
+
+export function decorateDNodes(dNodes: DNode[], predicate: DecoratorPredicate, modifier: DecoratorModifier): DNode[] {
+	let nodes = [ ...dNodes ];
+	while (nodes.length) {
+		const node = nodes.pop();
+		if (node && typeof node !== 'string') {
+			if (predicate(node)) {
+				modifier(node);
+			}
+			if (node.children) {
+				nodes = [ ...nodes, ...(<any> node).children ];
+			}
+		}
+	}
+	return dNodes;
+}
+
 export const registry = new FactoryRegistry();
 
 export function w<P extends WidgetProperties>(factory: WidgetFactory<Widget<P>, P> | string, properties: P): WNode;
